requirements_dev.txt
wheel
watchdog
flake8
tox
coverage
Sphinx
twine
black


# dependencies for the project

requests
celery
django
djangorestframework
django-environ
django-filter
openai
jsonschema